requirements.txt
# FastAPI 및 웹 서버
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OpenAI API
openai>=1.3.0

# 이미지 처리
opencv-python>=4.8.0
Pillow>=10.0.0

# 데이터 처리
pandas>=2.1.0
numpy>=1.24.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 성능 평가 (선택적)
scikit-learn>=1.3.0

# 데이터 검증
pydantic>=2.0.0

